nixos/testing: Add meta attribute to test runners.
Instead of adding the meta attribute only to the final return value of
makeTest, we now add the meta attribute to the test/coverage runner
directly and thus get the attribute even if the .test attribute is used.

Previously, meta attributes were passed only the resulting derivation
attributes from makeTest, but not the attribute "test" within those
derivation attributes.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>